+++ CHANGELOG.md
@@ -33,7 +33,8 @@ Changelog
      `getLastModifiedAt()` to `FileResourceInterface`
    * removed all pattern-related classes. This logic is now provided by the
      `Selector` class
-   * `ResourceRepository::remove()` now returns the number of removed resources
+   * `ResourceRepository::remove()`, `tag()` and `untag()` now return the number
+     of affected resources
 
 * 1.0.0-alpha3 (2014-02-22)